+++ packages/helia/src/utils/libp2p-defaults.ts
@@ -2,7 +2,7 @@ import { noise } from '@chainsafe/libp2p-noise'
 import { yamux } from '@chainsafe/libp2p-yamux'
 import { createDelegatedRoutingV1HttpApiClient } from '@helia/delegated-routing-v1-http-api-client'
 import { delegatedHTTPRoutingDefaults } from '@helia/routers'
-import { autoTLS } from '@libp2p/auto-tls'
+import { autoTLS } from '@ipshipyard/libp2p-auto-tls'
 import { autoNAT } from '@libp2p/autonat'
 import { bootstrap } from '@libp2p/bootstrap'
 import { circuitRelayTransport, circuitRelayServer, type CircuitRelayService } from '@libp2p/circuit-relay-v2'
@@ -20,11 +20,11 @@ import { webRTC, webRTCDirect } from '@libp2p/webrtc'
 import { webSockets } from '@libp2p/websockets'
 import { ipnsSelector } from 'ipns/selector'
 import { ipnsValidator } from 'ipns/validator'
-import * as libp2pInfo from 'libp2p/version'
+import { userAgent } from 'libp2p/user-agent'
 import { name, version } from '../version.js'
 import { bootstrapConfig } from './bootstrappers.js'
 import type { Libp2pDefaultsOptions } from './libp2p.js'
-import type { AutoTLS } from '@libp2p/auto-tls'
+import type { AutoTLS } from '@ipshipyard/libp2p-auto-tls'
 import type { Libp2pOptions } from 'libp2p'
 
 export interface DefaultLibp2pServices extends Record<string, unknown> {
@@ -41,19 +41,23 @@ export interface DefaultLibp2pServices extends Record<string, unknown> {
 }
 
 export function libp2pDefaults (options: Libp2pDefaultsOptions = {}): Libp2pOptions<DefaultLibp2pServices> & Required<Pick<Libp2pOptions<DefaultLibp2pServices>, 'services'>> {
-  const agentVersion = `${name}/${version} ${libp2pInfo.name}/${libp2pInfo.version} UserAgent=${process.version}`
+  const agentVersion = `${name}/${version} ${userAgent()}`
 
   return {
     privateKey: options.privateKey,
     dns: options.dns,
+    nodeInfo: {
+      userAgent: agentVersion
+    },
     addresses: {
       listen: [
         '/ip4/0.0.0.0/tcp/0',
         '/ip4/0.0.0.0/tcp/0/ws',
+        '/ip4/0.0.0.0/udp/0/webrtc-direct',
         '/ip6/::/tcp/0',
         '/ip6/::/tcp/0/ws',
-        '/p2p-circuit',
-        '/webrtc'
+        '/ip6/::/udp/0/webrtc-direct',
+        '/p2p-circuit'
       ]
     },
     transports: [
@@ -88,12 +92,8 @@ export function libp2pDefaults (options: Libp2pDefaultsOptions = {}): Libp2pOpti
           ipns: ipnsSelector
         }
       }),
-      identify: identify({
-        agentVersion
-      }),
-      identifyPush: identifyPush({
-        agentVersion
-      }),
+      identify: identify(),
+      identifyPush: identifyPush(),
       keychain: keychain(options.keychain),
       ping: ping(),
       relay: circuitRelayServer(),