+++ packages/dag-cbor/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [4.0.3](https://github.com/ipfs/helia/compare/dag-cbor-v4.0.2...dag-cbor-v4.0.3) (2025-03-13)
+
+
+### Documentation
+
+* add spell checker to ci ([#743](https://github.com/ipfs/helia/issues/743)) ([45ca6bc](https://github.com/ipfs/helia/commit/45ca6bc70b1644028500101044595fa0e2199b07))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^5.2.0 to ^5.2.1
+
+## [4.0.2](https://github.com/ipfs/helia/compare/dag-cbor-v4.0.1...dag-cbor-v4.0.2) (2025-01-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @helia/interface bumped from ^5.1.0 to ^5.2.0
+
 ## [4.0.1](https://github.com/ipfs/helia/compare/dag-cbor-v4.0.0...dag-cbor-v4.0.1) (2024-11-18)